Move mapPanels outside components
