+++ app/src/components/MapContainer.jsx
@@ -7,17 +7,17 @@ import delay from 'lodash/delay';
 import template from 'lodash/template';
 import templateSettings from 'lodash/templateSettings';
 import { MIN_ZOOM_LEVEL } from 'config';
-import ControlPanel from 'components/MapPanels/rightControlPanel/containers/ControlPanel';
+import ControlPanel from 'mapPanels/rightControlPanel/containers/ControlPanel';
 import Timebar from 'timebar/containers/Timebar';
-import Leyend from 'components/MapPanels/Leyend';
-import MiniGlobe from 'components/MapPanels/MiniGlobe';
-import MobileLeftExpand from 'components/MapPanels/MobileLeftExpand';
+import Leyend from 'mapPanels/leftControlPanel/components/Leyend';
+import MiniGlobe from 'mapPanels/leftControlPanel/components/MiniGlobe';
+import MobileLeftExpand from 'mapPanels/leftControlPanel/components/MobileLeftExpand';
 import ReportPanel from 'report/containers/ReportPanel';
 import MapLayers from 'containers/Layers/MapLayers';
 import DrawingManager from 'containers/Map/DrawingManager';
 import Areas from 'areasOfInterest/containers/Areas';
 import MapFooter from 'siteNav/components/MapFooter';
-import LeftControlPanel from 'components/MapPanels/leftControlPanel/containers/LeftControlPanel';
+import LeftControlPanel from 'mapPanels/leftControlPanel/containers/LeftControlPanel';
 import mapStyles from 'styles/components/map.scss';
 import mapPanelsStyles from 'styles/components/map-panels.scss';
 import Attributions from 'siteNav/components/Attributions';
