+++ app/src/mapPanels/leftControlPanel/components/Leyend.jsx
