+++ app/src/mapPanels/leftControlPanel/components/Loader.jsx
