+++ app/src/mapPanels/leftControlPanel/components/MiniGlobe.jsx
