+++ app/src/mapPanels/leftControlPanel/components/ZoomControls.jsx
