+++ app/src/mapPanels/leftControlPanel/containers/Loader.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import Loader from 'components/MapPanels/leftControlPanel/components/Loader';
+import Loader from 'mapPanels/leftControlPanel/components/Loader';
 
 const mapStateToProps = state => ({
   visible: state.map.loading || (state.vesselInfo.infoPanelStatus && state.vesselInfo.infoPanelStatus.isLoading)
