+++ app/src/mapPanels/rightControlPanel/components/ControlPanel.jsx
@@ -4,8 +4,8 @@ import classnames from 'classnames';
 import { CONTROL_PANEL_MENUS } from 'constants';
 import MediaQuery from 'react-responsive';
 import AreasPanel from 'areasOfInterest/containers/AreasPanel';
-import MenuLink from 'components/MapPanels/rightControlPanel/components/MenuLink';
-import SubMenu from 'components/MapPanels/rightControlPanel/containers/SubMenu';
+import MenuLink from 'mapPanels/rightControlPanel/components/MenuLink';
+import SubMenu from 'mapPanels/rightControlPanel/containers/SubMenu';
 import FilterPanel from 'filters/containers/FilterPanel';
 import FilterGroupPanel from 'filters/containers/FilterGroupPanel';
 import LayerPanel from 'layers/containers/LayerPanel';
