+++ app/src/mapPanels/rightControlPanel/containers/SubMenu.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
-import SubMenu from 'components/MapPanels/rightControlPanel/components/SubMenu';
-import { setSubmenu } from 'components/MapPanels/rightControlPanel/rightControlPanelActions';
+import SubMenu from 'mapPanels/rightControlPanel/components/SubMenu';
+import { setSubmenu } from 'mapPanels/rightControlPanel/rightControlPanelActions';
 
 
 const mapDispatchToProps = dispatch => ({
