+++ app/src/mapPanels/rightControlPanel/rightControlPanelActions.js
